Moving fulltext parsing functions to common
