+++ pdf_extraction/common/utilities.py
@@ -100,12 +100,19 @@ def extract_text_from_pdf(pdf_path):
 #Preprocess Text
 def preprocess_text(text):
     # Remove References/Bibliography and Acknowledgements sections
-    text = re.sub(r'\bREFERENCES\b.*', '', text, flags=re.DOTALL | re.IGNORECASE)
-    text = re.sub(r'\bACKNOWLEDGEMENTS\b.*', '', text, flags=re.DOTALL | re.IGNORECASE)
-    text = re.sub(r'\bBIBLIOGRAPHY\b.*', '', text, flags=re.DOTALL | re.IGNORECASE)
-    # Tokenize text into sentences
-    sentences = sent_tokenize(text)
-    return sentences
+	text = re.sub(r'\bREFERENCES\b.*', '', text, flags=re.DOTALL | re.IGNORECASE)
+	text = re.sub(r'\bACKNOWLEDGEMENTS\b.*', '', text, flags=re.DOTALL | re.IGNORECASE)
+	text = re.sub(r'\bBIBLIOGRAPHY\b.*', '', text, flags=re.DOTALL | re.IGNORECASE)
+
+	# Remove hyphenation at line breaks and join split words
+	text = re.sub(r'-\n', '', text)
+
+	# Normalize whitespace (remove multiple spaces and trim)
+	#text = re.sub(r'\s+', ' ', text).strip()
+
+	# Tokenize text into sentences
+	sentences = sent_tokenize(text)
+	return sentences
 
 def identify_sections(sentences, section_mapping):
     sections = {'abstract','introduction','methods','results','discussion' }
@@ -128,6 +135,11 @@ def identify_sections(sentences, section_mapping):
             sections[current_section].append(sentence)
     return sections
 
+#The output is much higher quality if we only focus on sentences which have 
+#been labeled as containing RESPONSE variables. 
+def filter_sentences(sentences, keywords):
+    filtered_sentences = [sentence for sentence in sentences if any(keyword in sentence.lower() for keyword in keywords)]
+    return filtered_sentences
 
 #==============================================================================
 # Functions related to spacy  
@@ -165,8 +177,6 @@ def clean_annotations(data):
 #This function will return the text and the entities for processing
 def extract_entities(text):
 	doc = nlp(text)
-	#This line is for extracting entities only
-	#entities = [(ent.text, ent.label_) for ent in doc.ents]
 	#This line is for extracting entities with dependencies. 
 	entities = [(ent.text, ent.label_, ent.start, ent.end) for ent in doc.ents]
 	return doc,entities
@@ -200,6 +210,156 @@ def find_entity_groups(doc, entities, label_type):
 			entity_groups.extend(sent_entity_groups)
 	# Removing duplicates and returning the result
 	return list(set(entity_groups))
+
+#==============================================================================
+# Use the output of a spacy NER to trace syntactical dependencies and 
+# build tables.  
+#==============================================================================
+#Function just to find ancestors of a token. 
+def get_ancestors(token):
+    ancestors = []
+    while token.head != token:
+        ancestors.append(token.head)
+        token = token.head
+    return ancestors
+
+# Function to find shortest path between two tokens in the
+# dependency tree based on the distance to a common ancestor
+# (least common ancestor, LCA)
+def find_shortest_path(token1, token2):
+    ancestors1 = get_ancestors(token1)
+    ancestors2 = get_ancestors(token2)
+    ancestors2.insert(0,token2)
+    #print(f"Ancestors 1 {ancestors1}")
+    #print(f"Ancestors 2 {ancestors2}")
+    # Find the lowest common ancestor
+    common_ancestor = None
+    for ancestor in ancestors1:
+        if ancestor in ancestors2:
+            common_ancestor = ancestor
+            break
+    if common_ancestor is None:
+        return float('inf')
+    # Calculate the distance as the number of nodes in the dependency tree
+    #print(f"Common ancestor {common_ancestor}")
+    distance1 = ancestors1.index(common_ancestor) + 1
+    distance2 = ancestors2.index(common_ancestor) + 1
+    #print(f"Distance1 = {distance1} and Distance2 = {distance2}")
+    distance = distance1 + distance2
+    return distance
+
+#Function to trace syntactical dependency back to a specific label
+#Use this to find the TREATMENT corresponding to a CARDINAL or PERCENTAGE
+#If you want to see the tree for a specific token use the print_tree
+#function defined below.
+def find_label_in_tree(token, label_id):
+	vnames = []
+	level = 0
+	for ancestor in token.ancestors:
+		print(f"ancestor: {ancestor}, all ancestors {list(token.ancestors)}")
+		for child in ancestor.children:
+			print(f"child: {child}, , all children {list(ancestor.children)}")
+			if child.ent_type_ in label_id:
+				vname = child.text.strip(',')
+				vnames.append(vname)
+				print(f"Names so far: {vnames}")
+			elif child.dep_ in ['nmod','nummod','conj', 'appos']:
+				print(f"Else if, next tree: {ancestor}")
+				find_label_in_tree(ancestor, label_id)
+		level += 1
+		print(f"level {level}")
+	return vnames
+
+
+#Function using heuristics to guess whether a sentence might actually be 
+#a table extracted as a single sentence. Since parsing text is messy, this 
+#provides a tool to infer the quality of output coming from a document. 
+def from_table(sent):
+	"""
+	Determine if a given sentence is likely from a table based on heuristic checks.
+	
+    Args:
+        sent: A spaCy Span object representing the sentence.
+
+    Returns:
+        bool: True if the sentence is likely from a table, False otherwise.
+	"""
+	text = sent.text
+	howtrue = 0 #Make this a scale from 0 to MAX
+	# Heuristic 1: Check for white space characters used in formatting
+	spaces = text.count('\u2009')
+	if(spaces>2):
+		howtrue +=1
+	# Heuristic 2: Check for consistent alignment/spacing
+	lines = text.split('\n')
+	if len(lines) > 1:
+		line_lengths = [len(line.strip()) for line in lines]
+		if max(line_lengths) - min(line_lengths) < 10:  # Threshold for alignment
+			howtrue += 1
+	# # Heuristic 3: Check for many newlines denoting tabular format
+	tabs = text.count('\n')
+	if(tabs > 10):
+		howtrue += 1
+	return(howtrue)
+
+# Function to create a table of treatments and responses using syntactical
+# dependencies within the sentence to infer how numbers and treatments are 
+# related. 
+def create_table(doc, entities, study_id):
+	data = []
+	responses = ['dry weight', 'biomass']
+	label_id = ["TREATMENT", "INOCTYPE"]
+	for response in responses:
+		response_ents = [ent for ent in entities if ent[1] == 'RESPONSE' and response in ent[0].lower()]
+		for resp_ent in response_ents:
+			resp_span = doc[resp_ent[2]:resp_ent[3]]
+			entities2 = [ent for ent in resp_span.sent.ents if ent.label_ in label_id]
+			for token in resp_span.root.head.subtree:
+				#Check it's a type we want, and not punctuation
+				if token.ent_type_ in ['CARDINAL', 'PERCENTAGE'] and token.text not in ['%', ' ', ',']:
+					value = token.text
+					ent1 = next((ent for ent in resp_span.sent.ents if token in ent), None)
+					#Find the connected treatment by parsing dependencies
+					shortest_distance = float('inf')
+					treat = None
+					for ent2 in entities2:
+						distance = find_shortest_path(ent1.root, ent2.root)
+						distance2 = abs(ent2.root.i)
+						#Handle the case of equal distances separately
+						if distance < shortest_distance:
+							shortest_distance = distance
+							shortest_distance2 = distance2
+							treat = ent2
+							#print(f"{treat}, {shortest_distance}")
+						#If dependence distances are equal, use whichever precedes the number
+						elif distance == shortest_distance:
+							if distance2 < shortest_distance2:
+								shortest_distance = distance
+								shortest_distance2 = distance2
+								treat = ent2
+					if token.ent_type_ == 'CARDINAL':
+						data.append({
+							'STUDY': study_id,
+							'TREATMENT': treat,
+							'RESPONSE': response,
+							'CARDINAL': value,
+							'PERCENTAGE': '',
+							'SENTENCE': token.sent,
+							'ISTABLE': from_table(token.sent)
+						})
+					elif token.ent_type_ == 'PERCENTAGE':
+						data.append({
+							'STUDY':study_id,
+							'TREATMENT': treat,
+							'RESPONSE': response,
+							'CARDINAL': '',
+							'PERCENTAGE': value,
+							'SENTENCE': token.sent,
+							'ISTABLE': from_table(token.sent)
+						})
+	#df = pd.DataFrame(data)
+	return data
+
 
 #==============================================================================
 # Functions for dealing with label studio
